Move color_log_text to ili2dbutils
