+++ projectgenerator/gui/import_data.py
@@ -25,6 +25,7 @@ from projectgenerator.gui.options import OptionsDialog, CompletionLineEdit
 from projectgenerator.gui.multiple_models import MultipleModelsDialog
 from projectgenerator.libili2db.iliimporter import JavaNotFoundError
 from projectgenerator.libili2db.ilicache import IliCache, ModelCompleterDelegate
+from projectgenerator.libili2db.ili2dbutils import color_log_text
 from projectgenerator.libqgsprojectgen.dbconnector import pg_connector
 from projectgenerator.utils.qt_utils import (
     make_file_selector,
@@ -33,8 +34,7 @@ from projectgenerator.utils.qt_utils import (
     Validators,
     FileValidator,
     NonEmptyStringValidator,
-    OverrideCursor,
-    color_log_text
+    OverrideCursor
 )
 from qgis.PyQt.QtGui import (
     QColor,
