+++ projectgenerator/libili2db/ili2dbutils.py
@@ -23,6 +23,7 @@ import zipfile
 import glob
 
 from qgis.PyQt.QtCore import QCoreApplication
+from qgis.PyQt.QtGui import QColor
 
 from projectgenerator.utils.qt_utils import download_file, NetworkError
 
@@ -87,3 +88,16 @@ def get_all_modeldir_in_path(path, lambdafunction=None):
                 lambdafunction(subdir)
             modeldir += subdir + ';'
     return modeldir[:-1]  # remove last ';'
+
+def color_log_text(text, txt_edit):
+    textlines = text.splitlines()
+    for textline in textlines:
+        if textline.startswith("Warning:"):
+            txt_edit.setTextColor(QColor('#FFBF00'))
+            txt_edit.append(textline)
+        elif "error" in textline.lower() or "failed" in textline.lower():
+            txt_edit.setTextColor(QColor('#aa2222'))
+            txt_edit.append(textline)
+        else:
+            txt_edit.setTextColor(QColor('#2a2a2a'))
+            txt_edit.append(textline)
